+++ praw/models/reddit/subreddit.py
@@ -9,6 +9,7 @@ from json import dumps, loads
 from os.path import basename, dirname, isfile, join
 from typing import TYPE_CHECKING, Any, Dict, Generator, Iterator, List, Optional, Union
 from urllib.parse import urljoin
+from warnings import warn
 from xml.etree.ElementTree import XML
 
 import websocket
@@ -2150,6 +2151,12 @@ class SubredditModeration:
     ) -> Iterator["praw.models.SubredditMessage"]:
         """Return a :class:`.ListingGenerator` for moderator messages.
 
+        .. warning::
+
+            Legacy modmail is being deprecated in June 2021. Please see
+            https://www.reddit.com/r/modnews/comments/mar9ha/even_more_modmail_improvements/
+            for more info.
+
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
 
@@ -2167,6 +2174,13 @@ class SubredditModeration:
                     print(f"From: {reply.author}, Body: {reply.body}")
 
         """
+        warn(
+            "Legacy modmail is being deprecated in June 2021. Please see"
+            " https://www.reddit.com/r/modnews/comments/mar9ha/even_more_modmail_improvements/"
+            " for more info.",
+            category=DeprecationWarning,
+            stacklevel=3,
+        )
         return ListingGenerator(
             self.subreddit._reddit,
             API_PATH["moderator_messages"].format(subreddit=self.subreddit),
@@ -2357,6 +2371,12 @@ class SubredditModeration:
     ) -> Iterator["praw.models.SubredditMessage"]:
         """Return a :class:`.ListingGenerator` for unread moderator messages.
 
+        .. warning::
+
+            Legacy modmail is being deprecated in June 2021. Please see
+            https://www.reddit.com/r/modnews/comments/mar9ha/even_more_modmail_improvements/
+            for more info.
+
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
 
@@ -2372,6 +2392,13 @@ class SubredditModeration:
                 print(f"From: {message.author}, To: {message.dest}")
 
         """
+        warn(
+            "Legacy modmail is being deprecated in June 2021. Please see"
+            " https://www.reddit.com/r/modnews/comments/mar9ha/even_more_modmail_improvements/"
+            " for more info.",
+            category=DeprecationWarning,
+            stacklevel=3,
+        )
         return ListingGenerator(
             self.subreddit._reddit,
             API_PATH["moderator_unread"].format(subreddit=self.subreddit),